fix(clippy): mechanical fixes batch 3 — 53 warnings
Fix push_str single-char, unnecessary mut, useless format!,
unneeded return, clone on Copy, collapsed if, match→if let,
ref-deref, len comparison, map_or simplification, split_once,
strip_prefix, range pattern, unused vars, and more.

+++ src/interpreter/learnable.rs
@@ -144,7 +144,7 @@ impl Interpreter {
                                     0.0
                                 }
                             } else {
-                                cosine_similarity(&query_embedding, &entry.embedding) as f32
+                                cosine_similarity(&query_embedding, &entry.embedding)
                                     * entry.confidence as f32
                                     * entry.priority as f32
                             };
